UserController 수정

+++ controller/UserController.js
@@ -3,6 +3,11 @@ const UserService = require("../services/UserService");
 module.exports = {
   getUserInfo: async (req, res) => {
     const { user_id } = req.params;
+    if (!user_id) {
+      return res.status(400).json({
+        message: "user_id가 존재하지 않습니다.",
+      });
+    }
     try {
       const result = await UserService.getUserInfo(user_id);
       return res.json(result[0]);
@@ -11,8 +16,29 @@ module.exports = {
     }
   },
   createUser: async (req, res) => {
+    const { account, password, name } = req.body;
+    if (!account) {
+      return res.status(400).json({
+        message: "account를 입력해주세요.",
+      });
+    }
+    if (!password) {
+      return res.status(400).json({
+        message: "password를 입력해주세요.",
+      });
+    }
+    if (password.length < 8 || password.length > 16) {
+      return res.statsu(400).json({
+        message: "password는 8~16자 사이여야 합니다.",
+      });
+    }
+    if (!name) {
+      return res.status(400).json({
+        message: "name을 입력해주세요.",
+      });
+    }
+
     try {
-      const { account, password, name } = req.body;
       const result = await UserService.createUser(account, password, name);
       return res.status(201).json({
         message: "회원 가입을 성공하였습니다.",
@@ -22,8 +48,13 @@ module.exports = {
     }
   },
   loginRequest: async (req, res) => {
+    const { account, password } = req.body;
+    if (!account) {
+      return res.status(400).json({
+        message: "account를 입력해주세요.",
+      });
+    }
     try {
-      const { account, password } = req.body;
       const result = await UserService.loginUser(account, password);
       if (result) {
         res.status(200).json({
@@ -39,8 +70,23 @@ module.exports = {
     }
   },
   updateUser: async (req, res) => {
+    const { password, name } = req.body;
+    if (!password) {
+      return res.status(400).json({
+        message: "password를 입력해주세요.",
+      });
+    }
+    if (password.length < 8 || password.length > 16) {
+      return res.statsu(400).json({
+        message: "password는 8~16자 사이여야 합니다.",
+      });
+    }
+    if (!name) {
+      return res.status(400).json({
+        message: "name을 입력해주세요.",
+      });
+    }
     try {
-      const { password, name } = req.body;
       const result = await UserService.updateUser(password, name);
       return res.status(200).json({
         message: "회원 정보 수정을 완료하였습니다.",
@@ -50,12 +96,39 @@ module.exports = {
     }
   },
   deleteUser: async (req, res) => {
+    const { user_id } = req.params;
+    if (!user_id) {
+      return res.status(400).json({
+        message: "user_id가 존재하지 않습니다.",
+      });
+    }
     try {
-      const { user_id } = req.params;
       const result = await UserService.deleteUser(user_id);
       return res.status(200).json({
         message: "회원 정보 삭제를 완료하였습니다.",
       });
+    } catch (err) {
+      throw err;
+    }
+  },
+  validationAccount: async (req, res) => {
+    const { account } = req.body;
+    if (!account) {
+      return res.status(400).json({
+        message: "account를 입력해주세요.",
+      });
+    }
+    try {
+      const result = await UserService.validationAccount(account);
+      if (result) {
+        return res.status(200).json({
+          message: "사용 가능한 아이디입니다.",
+        });
+      } else {
+        return res.status(400).json({
+          message: "이미 사용중인 아이디입니다.",
+        });
+      }
     } catch (err) {
       throw err;
     }
